pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "page analyzer"
authors = ["dmitry_rodin <dimoon-96@yandex.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
flake8 = "^7.0.0"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.28.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
datetime = "^5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
